backend/requirements.txt
openai
python-dotenv
google-cloud-speech
pymongo
werkzeug
flask
flask-cors
backend
